+++ products/fire.browser.js
@@ -41,9 +41,6 @@
       )
     }
     async execute() {
-      return this.executeSync()
-    }
-    executeSync() {
       let outputLines = []
       const _originalConsoleLog = console.log
       const tempConsoleLog = (...params) => outputLines.push(params)
@@ -107,9 +104,6 @@ fireNode
  catchAllNodeType errorNode
  javascript
   async execute() {
-   return this.executeSync()
-  }
-  executeSync() {
    let outputLines = []
    const _originalConsoleLog = console.log
    const tempConsoleLog = (...params) => outputLines.push(params)