+++ products/jtree.browser.js
@@ -835,9 +835,7 @@ class TreeNode extends AbstractNode {
     this._setLine(line)
     this._setChildren(children)
   }
-  execute(context) {
-    return Promise.all(this.map(child => child.execute(context)))
-  }
+  execute() {}
   async loadRequirements(context) {
     await Promise.all(this.map(node => node.loadRequirements(context)))
   }
@@ -847,9 +845,6 @@ class TreeNode extends AbstractNode {
   getLineCellTypes() {
     // todo: make this any a constant
     return "undefinedCellType ".repeat(this.getWords().length).trim()
-  }
-  executeSync(context) {
-    return this.map(child => child.executeSync(context))
   }
   isNodeJs() {
     return typeof exports !== "undefined"